separated files for assembly into pentomino.py;    functional sides.py;    functional numext.py with supervision
